part 4 finished!

+++ osa4/Tehtävät/blogilista/blogiBack/controllers/blogit.js
@@ -1,6 +1,7 @@
 const blogRouter = require("express").Router()
 const Blog = require("../models/blogi")
-
+const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
 const User = require("../models/user")
 
 blogRouter.get("/", async (request, response) => {
@@ -18,9 +19,13 @@ blogRouter.get("/:id", async (request, response, next) => {
 	}
 })
 
-blogRouter.post("/", async (request, response, next) => {
+blogRouter.post("/", middleware.userExtractor,async (request, response, next) => {
 	const body = request.body
-	const user = await User.findById(body.userId)
+	const decodedToken = jwt.verify(request.token, process.env.SECRET)
+	if (!request.token || !decodedToken.id) {
+		return response.status(401).json({ error: 'token missing or invalid' })
+	}
+	const user = request.user
 
 	const newBlog = new Blog({
 		title: body.title,
@@ -44,8 +49,21 @@ blogRouter.post("/", async (request, response, next) => {
 })
 
 blogRouter.delete("/:id", async (request, response, next) => {
-	await Blog.findByIdAndRemove(request.params.id)
-	response.status(204).end()
+	const decodedToken = jwt.verify(request.token, process.env.SECRET)
+	if (!request.token || !decodedToken.id) {
+		return response.status(401).json({ error: 'token missing or invalid' })
+	}
+	const user = await User.findById(decodedToken.id)
+	const blogToRemove = await Blog.findById(request.params.id)
+	console.log(blogToRemove)
+	if (user.id = blogToRemove.user)
+	{
+		const removedBlog = await Blog.findByIdAndRemove(request.params.id)
+		response.status(204).json(removedBlog.toJSON())
+	}
+	else{
+	response.status(401).end()
+	}
 })
 
 blogRouter.put("/:id", async (request, response, next) => {
